Move ServerApi to its own file for g3 BUILD configuration.
